requirements.txt
Django
djangorestframework
django-cors-headers
psutil
dotenv
gunicorn
apiMoneyFusion
